requirements.txt
pandas
requests
python-dotenv
asyncpraw
beautifulsoup4
fastapi
uvicorn